+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 *UNRELEASED*
 
 * Added `Field\CompositeTypeResolver` to make composing type resolvers easier
+* Changed constructors of `Entity\EntityResolver` and `Field\FieldResolver` to
+  accept a callback to allow for inclusion or exclusion of entities and fields
+  instead of an array only allowing for exclusions
 
 ## [0.1.0](https://github.com/elazar/phanua/releases/tag/0.1.0)